refactor : 컴포넌트 렌더링 함수 분리
