+++ src/component/AddLunchModal.js
@@ -1,0 +1,13 @@
+import { DOM } from "../utils/dom.js";
+
+const AddLunchModal = {
+  open() {
+    DOM.$modal.classList.add("modal--open");
+  },
+
+  close() {
+    DOM.$modal.classList.remove("modal--open");
+  },
+};
+
+export default AddLunchModal;
