--- src/component/AddLunchModal/elements.js
@@ -1,3 +1,0 @@
-import { $ } from "../../utils/querySelectors.js";
-
-export const $modal = $(".modal");
