+++ src/utils/render.js
@@ -1,0 +1,7 @@
+import { DOM } from "./dom.js";
+
+function render(element, dom) {
+  dom.append(element);
+}
+
+export default render;
